Add platform-specific logic
