+++ src/platforms/browser/get-token.js
@@ -1,0 +1,1 @@
+module.exports = function () {}
